feat(Status.java): 폰이 같은 열에 있을 때 다른 점수로 계산하는 로직 구현
탄은 팀의 폰이 같은 열에 있다면, 해당 폰들의 점수를 0.5로 계산하는 로직
구현

+++ src/main/java/chess/domain/PieceKind.java
@@ -22,7 +22,7 @@ public enum PieceKind {
         return piece.isSamePiece(pieceName);
     }
 
-    public double calculateScore(final int count) {
+    public double calculateScore(final double count) {
         return score * count;
     }
 }
